Add suport for now.js config file
